requirements.txt
djangorestframework_simplejwt==4.8.0
Django==3.2.7
djangorestframework==3.12.4
